requirements.txt
Django==1.11.4
django-cleanup==1.0.1
django-dbbackup==3.2.0
django-payu==0.5
lml==0.0.1
MySQL-python==1.2.5
olefile==0.44
Pillow==4.2.1
pkg-resources==0.0.0
pytz==2017.2
six==1.11.0
texttable==1.1.1
webencodings==0.5.1
